json-rpc: add a common json-rpc server builder
