+++ crates/sui-gateway/Cargo.toml
@@ -20,6 +20,7 @@ tracing = "0.1.34"
 tokio = { version = "1.18.2", features = ["full"] }
 futures = "0.3.21"
 ed25519-dalek = { version = "1.0.1", features = ["batch", "serde"] }
+prometheus_exporter = "0.8.4"
 
 sui_core = { path = "../../sui_core" }
 sui-config = { path = "../sui-config" }
