+++ crates/sui-gateway/src/lib.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod api;
+pub mod config;
+pub mod json_rpc;
 pub mod rpc_gateway;
 pub mod rpc_gateway_client;
-pub mod config;
